refactor: extract inline modals from TemplateSearchModal into separate files (#387)
- Extracted ExternalItemEditModal and ExternalBarCodeInsertModal from TemplateSearchModal.tsx into their own files for better organization and maintainability
- Created src/sections/search/components/ExternalTemplateToItemGroupModal.tsx and src/sections/search/components/ExternalBarCodeInsertModal.tsx
- Updated TemplateSearchModal.tsx to import and use the new components
- Cleaned up unused imports and removed inline component definitions
- Ensured all functionality and type safety remain intact

Closes #387
