+++ src/sections/search/components/ExternalTemplateToItemGroupModal.tsx
@@ -1,0 +1,75 @@
+import { type Accessor, type Setter } from 'solid-js'
+import toast from 'solid-toast'
+import { type Recipe } from '~/modules/diet/recipe/domain/recipe'
+import { ItemEditModal } from '~/sections/food-item/components/ItemEditModal'
+import {
+  type ItemGroup,
+  type RecipedItemGroup,
+  type SimpleItemGroup,
+  createSimpleItemGroup,
+  createRecipedItemGroup,
+} from '~/modules/diet/item-group/domain/itemGroup'
+import { type Template } from '~/modules/diet/template/domain/template'
+import { type TemplateItem } from '~/modules/diet/template-item/domain/templateItem'
+import { ModalContextProvider } from '~/sections/common/context/ModalContext'
+import { formatError } from '~/shared/formatError'
+
+export interface ExternalTemplateToItemGroupModalProps {
+  visible: Accessor<boolean>
+  setVisible: Setter<boolean>
+  selectedTemplate: Accessor<Template>
+  targetName: string
+  onNewItemGroup: (
+    newGroup: ItemGroup,
+    originalAddedItem: TemplateItem,
+  ) => Promise<void>
+}
+
+export function ExternalTemplateToItemGroupModal(props: ExternalTemplateToItemGroupModalProps) {
+  return (
+    <ModalContextProvider visible={props.visible} setVisible={props.setVisible}>
+      <ItemEditModal
+        targetName={props.targetName}
+        item={() => ({
+          reference: props.selectedTemplate().id,
+          name: props.selectedTemplate().name,
+          macros: props.selectedTemplate().macros,
+          __type:
+            props.selectedTemplate().__type === 'Food'
+              ? 'Item'
+              : 'RecipeItem', // TODO: Refactor conversion from template type to group/item types
+        })}
+        macroOverflow={() => ({
+          enable: true,
+        })}
+        onApply={(item) => {
+          // TODO: Refactor conversion from template type to group/item types
+          if (item.__type === 'Item') {
+            const newGroup: SimpleItemGroup = createSimpleItemGroup({
+              name: item.name,
+              items: [item],
+            })
+            props.onNewItemGroup(newGroup, item).catch((err) => {
+              console.error(err)
+              toast.error(
+                `Erro ao adicionar item: ${formatError(err)}`,
+              )
+            })
+          } else {
+            const newGroup: RecipedItemGroup = createRecipedItemGroup({
+              name: item.name,
+              recipe: (props.selectedTemplate() as Recipe).id,
+              items: [...(props.selectedTemplate() as Recipe).items],
+            })
+            props.onNewItemGroup(newGroup, item).catch((err) => {
+              console.error(err)
+              toast.error(
+                `Erro ao adicionar item: ${formatError(err)}`,
+              )
+            })
+          }
+        }}
+      />
+    </ModalContextProvider>
+  )
+}
